refactor(xlsx): extract types and styles modules from parser lib.rs
Split the 5804-line parser/src/lib.rs following the docx parser model
(separate modules per concern). No behavior change:
- types.rs: all data-model structs/enums (serde Serialize), ~1224 lines
- styles.rs: styles.xml parsing (fonts/fills/borders/cellXfs/dxfs/numFmts)

lib.rs drops from 5804 to 4220 lines. Shared helpers (parse_color,
read_zip_entry) are now pub(crate). Verified behavior-invariant via VRT
regression mode (65/65 pass, 0.0% diff vs pre-split baseline).
