+++ packages/xlsx/parser/src/styles.rs
@@ -1,0 +1,372 @@
+use crate::types::*;
+use crate::{parse_color, read_zip_entry};
+use std::io::Cursor;
+
+
+/// Resolve the workbook's Normal-style font (family name + point size) by
+/// following `<cellStyleXfs>[0].fontId` → `<fonts>[fontId]`. Returns `(None,
+/// None)` if `xl/styles.xml` is missing or malformed. The renderer uses this
+/// to compute the Max Digit Width for column-width pixel conversion
+/// (ECMA-376 §18.3.1.13).
+pub(crate) fn parse_default_font(archive: &mut zip::ZipArchive<Cursor<&[u8]>>) -> (Option<String>, Option<f64>) {
+    let Ok(xml) = read_zip_entry(archive, "xl/styles.xml") else { return (None, None); };
+    let Ok(doc) = roxmltree::Document::parse(&xml) else { return (None, None); };
+    let ns = "http://schemas.openxmlformats.org/spreadsheetml/2006/main";
+    let mut font_id: usize = 0;
+    for n in doc.descendants() {
+        if n.tag_name().name() == "cellStyleXfs" && n.tag_name().namespace() == Some(ns) {
+            if let Some(xf) = n.children().find(|c| c.is_element() && c.tag_name().name() == "xf") {
+                font_id = xf.attribute("fontId").and_then(|s| s.parse().ok()).unwrap_or(0);
+            }
+            break;
+        }
+    }
+    for fonts_node in doc.descendants() {
+        if fonts_node.tag_name().name() != "fonts" || fonts_node.tag_name().namespace() != Some(ns) { continue; }
+        if let Some(font_node) = fonts_node.children()
+            .filter(|c| c.is_element() && c.tag_name().name() == "font")
+            .nth(font_id)
+        {
+            let mut name = None;
+            let mut sz = None;
+            for child in font_node.children() {
+                match child.tag_name().name() {
+                    "name" => name = child.attribute("val").map(|s| s.to_string()),
+                    "sz"   => sz = child.attribute("val").and_then(|s| s.parse().ok()),
+                    _ => {}
+                }
+            }
+            return (name, sz);
+        }
+        break;
+    }
+    (None, None)
+}
+
+pub(crate) fn parse_styles(archive: &mut zip::ZipArchive<Cursor<&[u8]>>, theme_colors: &[String]) -> Result<Styles, String> {
+    let xml = read_zip_entry(archive, "xl/styles.xml")?;
+    let doc = roxmltree::Document::parse(&xml).map_err(|e| e.to_string())?;
+    let ns = "http://schemas.openxmlformats.org/spreadsheetml/2006/main";
+
+    let num_fmts = parse_num_fmts(&doc, ns);
+    let fonts = parse_fonts(&doc, ns, theme_colors);
+    let fills = parse_fills(&doc, ns, theme_colors);
+    let borders = parse_borders(&doc, ns, theme_colors);
+    let cell_xfs = parse_cell_xfs(&doc, ns);
+    let dxfs = parse_dxfs(&doc, ns, theme_colors);
+
+    Ok(Styles { fonts, fills, borders, cell_xfs, num_fmts, dxfs })
+}
+
+pub(crate) fn parse_dxfs(doc: &roxmltree::Document, ns: &str, theme_colors: &[String]) -> Vec<Dxf> {
+    let mut dxfs = Vec::new();
+    for dxfs_node in doc.descendants() {
+        if dxfs_node.tag_name().name() != "dxfs" || dxfs_node.tag_name().namespace() != Some(ns) {
+            continue;
+        }
+        for dxf_node in dxfs_node.children() {
+            if dxf_node.tag_name().name() != "dxf" { continue; }
+            let mut d = Dxf::default();
+            for child in dxf_node.children() {
+                match child.tag_name().name() {
+                    "font" => {
+                        let mut f = Font { size: 11.0, ..Default::default() };
+                        for fc in child.children() {
+                            match fc.tag_name().name() {
+                                "b" => f.bold = parse_st_on_off(&fc),
+                                "i" => f.italic = parse_st_on_off(&fc),
+                                "u" => {
+                                    let v = fc.attribute("val").unwrap_or("single");
+                                    if v != "none" {
+                                        f.underline = true;
+                                        if v != "single" {
+                                            f.underline_style = Some(v.to_string());
+                                        }
+                                    }
+                                }
+                                "strike" => f.strike = parse_st_on_off(&fc),
+                                "vertAlign" => {
+                                    if let Some(v) = fc.attribute("val") {
+                                        if v != "baseline" {
+                                            f.vert_align = Some(v.to_string());
+                                        }
+                                    }
+                                }
+                                "sz" => {
+                                    if let Some(v) = fc.attribute("val").and_then(|s| s.parse().ok()) {
+                                        f.size = v;
+                                    }
+                                }
+                                "name" => {
+                                    f.name = fc.attribute("val").map(|s| s.to_string());
+                                }
+                                "color" => {
+                                    f.color = parse_color(&fc, theme_colors);
+                                }
+                                _ => {}
+                            }
+                        }
+                        d.font = Some(f);
+                    }
+                    "fill" => {
+                        let mut f = Fill::default();
+                        for pf in child.children() {
+                            if pf.tag_name().name() == "patternFill" {
+                                f.pattern_type = pf.attribute("patternType").unwrap_or("solid").to_string();
+                                for color_node in pf.children() {
+                                    match color_node.tag_name().name() {
+                                        "fgColor" => f.fg_color = parse_color(&color_node, theme_colors),
+                                        "bgColor" => f.bg_color = parse_color(&color_node, theme_colors),
+                                        _ => {}
+                                    }
+                                }
+                            }
+                        }
+                        // In dxf, conditional format fills often only have bgColor; mirror into fgColor
+                        if f.fg_color.is_none() && f.bg_color.is_some() {
+                            f.fg_color = f.bg_color.clone();
+                        }
+                        d.fill = Some(f);
+                    }
+                    "border" => {
+                        let mut b = Border::default();
+                        for edge_node in child.children() {
+                            let style = edge_node.attribute("style").unwrap_or("").to_string();
+                            if style.is_empty() { continue; }
+                            let color = edge_node.children().find(|c| c.is_element())
+                                .and_then(|c| parse_color(&c, theme_colors));
+                            let edge = Some(BorderEdge { style, color });
+                            match edge_node.tag_name().name() {
+                                "left" => b.left = edge,
+                                "right" => b.right = edge,
+                                "top" => b.top = edge,
+                                "bottom" => b.bottom = edge,
+                                "horizontal" => b.horizontal = edge,
+                                "vertical"   => b.vertical   = edge,
+                                _ => {}
+                            }
+                        }
+                        d.border = Some(b);
+                    }
+                    "numFmt" => {
+                        let num_fmt_id = child.attribute("numFmtId")
+                            .and_then(|v| v.parse().ok()).unwrap_or(0);
+                        let format_code = child.attribute("formatCode")
+                            .unwrap_or("").to_string();
+                        d.num_fmt = Some(NumFmt { num_fmt_id, format_code });
+                    }
+                    _ => {}
+                }
+            }
+            dxfs.push(d);
+        }
+        break;
+    }
+    dxfs
+}
+
+pub(crate) fn parse_num_fmts(doc: &roxmltree::Document, ns: &str) -> Vec<NumFmt> {
+    let mut fmts = Vec::new();
+    for node in doc.descendants() {
+        if node.tag_name().name() == "numFmts" && node.tag_name().namespace() == Some(ns) {
+            for child in node.children() {
+                if child.tag_name().name() != "numFmt" { continue; }
+                let num_fmt_id = child.attribute("numFmtId").and_then(|v| v.parse().ok()).unwrap_or(0);
+                let format_code = child.attribute("formatCode").unwrap_or("").to_string();
+                fmts.push(NumFmt { num_fmt_id, format_code });
+            }
+            break;
+        }
+    }
+    fmts
+}
+
+/// ECMA-376 §22.9.2 ST_OnOff. Toggle elements like `<b/>`, `<i/>`, `<strike/>`
+/// accept an optional `val` attribute whose value is "1" / "true" (on) or
+/// "0" / "false" (off). When omitted, the presence of the element itself
+/// implies "on". A dxf with `<i val="0"/>` therefore means "differential
+/// format that *clears* italic", not "set italic to true".
+pub(crate) fn parse_st_on_off(node: &roxmltree::Node) -> bool {
+    match node.attribute("val") {
+        None => true,
+        Some(v) => !matches!(v, "0" | "false" | "False" | "FALSE" | "off" | "Off"),
+    }
+}
+
+pub(crate) fn parse_fonts(doc: &roxmltree::Document, ns: &str, theme_colors: &[String]) -> Vec<Font> {
+    let mut fonts = Vec::new();
+    for fonts_node in doc.descendants() {
+        if fonts_node.tag_name().name() == "fonts" && fonts_node.tag_name().namespace() == Some(ns) {
+            for font_node in fonts_node.children() {
+                if font_node.tag_name().name() != "font" { continue; }
+                let mut f = Font { size: 11.0, ..Default::default() };
+                for child in font_node.children() {
+                    match child.tag_name().name() {
+                        "b" => f.bold = parse_st_on_off(&child),
+                        "i" => f.italic = parse_st_on_off(&child),
+                        "u" => {
+                            let v = child.attribute("val").unwrap_or("single");
+                            if v != "none" {
+                                f.underline = true;
+                                if v != "single" {
+                                    f.underline_style = Some(v.to_string());
+                                }
+                            }
+                        }
+                        "strike" => f.strike = parse_st_on_off(&child),
+                        "vertAlign" => {
+                            if let Some(v) = child.attribute("val") {
+                                if v != "baseline" {
+                                    f.vert_align = Some(v.to_string());
+                                }
+                            }
+                        }
+                        "sz" => {
+                            if let Some(v) = child.attribute("val").and_then(|s| s.parse().ok()) {
+                                f.size = v;
+                            }
+                        }
+                        "name" => {
+                            f.name = child.attribute("val").map(|s| s.to_string());
+                        }
+                        "color" => {
+                            f.color = parse_color(&child, theme_colors);
+                        }
+                        _ => {}
+                    }
+                }
+                fonts.push(f);
+            }
+            break;
+        }
+    }
+    fonts
+}
+
+pub(crate) fn parse_fills(doc: &roxmltree::Document, ns: &str, theme_colors: &[String]) -> Vec<Fill> {
+    let mut fills = Vec::new();
+    for fills_node in doc.descendants() {
+        if fills_node.tag_name().name() == "fills" && fills_node.tag_name().namespace() == Some(ns) {
+            for fill_node in fills_node.children() {
+                if fill_node.tag_name().name() != "fill" { continue; }
+                let mut f = Fill::default();
+                for pf in fill_node.children() {
+                    match pf.tag_name().name() {
+                        "patternFill" => {
+                            f.pattern_type = pf.attribute("patternType").unwrap_or("none").to_string();
+                            for color_node in pf.children() {
+                                match color_node.tag_name().name() {
+                                    "fgColor" => f.fg_color = parse_color(&color_node, theme_colors),
+                                    "bgColor" => f.bg_color = parse_color(&color_node, theme_colors),
+                                    _ => {}
+                                }
+                            }
+                        }
+                        "gradientFill" => {
+                            // ECMA-376 §18.8.24 gradientFill — linear (default) uses
+                            // `degree`, path uses top/bottom/left/right as a relative
+                            // bounding box; children <stop position="n"><color/></stop>.
+                            let gtype = pf.attribute("type").unwrap_or("linear").to_string();
+                            let degree = pf.attribute("degree").and_then(|s| s.parse::<f64>().ok()).unwrap_or(0.0);
+                            let left   = pf.attribute("left").and_then(|s| s.parse::<f64>().ok()).unwrap_or(0.0);
+                            let right  = pf.attribute("right").and_then(|s| s.parse::<f64>().ok()).unwrap_or(0.0);
+                            let top    = pf.attribute("top").and_then(|s| s.parse::<f64>().ok()).unwrap_or(0.0);
+                            let bottom = pf.attribute("bottom").and_then(|s| s.parse::<f64>().ok()).unwrap_or(0.0);
+                            let mut stops: Vec<GradientStopSpec> = pf.children()
+                                .filter(|n| n.is_element() && n.tag_name().name() == "stop")
+                                .filter_map(|stop| {
+                                    let position = stop.attribute("position").and_then(|s| s.parse::<f64>().ok())?;
+                                    let color_node = stop.children().find(|c| c.is_element() && c.tag_name().name() == "color")?;
+                                    let color = parse_color(&color_node, theme_colors)?;
+                                    Some(GradientStopSpec { position, color })
+                                })
+                                .collect();
+                            stops.sort_by(|a, b| a.position.partial_cmp(&b.position).unwrap_or(std::cmp::Ordering::Equal));
+                            if !stops.is_empty() {
+                                f.gradient = Some(GradientFillSpec {
+                                    gradient_type: gtype,
+                                    degree,
+                                    left, right, top, bottom,
+                                    stops,
+                                });
+                            }
+                        }
+                        _ => {}
+                    }
+                }
+                fills.push(f);
+            }
+            break;
+        }
+    }
+    fills
+}
+
+pub(crate) fn parse_borders(doc: &roxmltree::Document, ns: &str, theme_colors: &[String]) -> Vec<Border> {
+    let mut borders = Vec::new();
+    for borders_node in doc.descendants() {
+        if borders_node.tag_name().name() == "borders" && borders_node.tag_name().namespace() == Some(ns) {
+            for border_node in borders_node.children() {
+                if border_node.tag_name().name() != "border" { continue; }
+                let has_diag_up = border_node.attribute("diagonalUp").map(|v| v == "1" || v == "true").unwrap_or(false);
+                let has_diag_down = border_node.attribute("diagonalDown").map(|v| v == "1" || v == "true").unwrap_or(false);
+                let mut b = Border::default();
+                let mut diag_edge: Option<BorderEdge> = None;
+                for edge_node in border_node.children() {
+                    let style = edge_node.attribute("style").unwrap_or("").to_string();
+                    let color = edge_node.children().find(|c| c.is_element()).and_then(|c| parse_color(&c, theme_colors));
+                    match edge_node.tag_name().name() {
+                        "left" if !style.is_empty() => b.left = Some(BorderEdge { style, color }),
+                        "right" if !style.is_empty() => b.right = Some(BorderEdge { style, color }),
+                        "top" if !style.is_empty() => b.top = Some(BorderEdge { style, color }),
+                        "bottom" if !style.is_empty() => b.bottom = Some(BorderEdge { style, color }),
+                        "diagonal" if !style.is_empty() => diag_edge = Some(BorderEdge { style, color }),
+                        _ => {}
+                    }
+                }
+                if has_diag_up { b.diagonal_up = diag_edge.clone(); }
+                if has_diag_down { b.diagonal_down = diag_edge; }
+                borders.push(b);
+            }
+            break;
+        }
+    }
+    borders
+}
+
+pub(crate) fn parse_cell_xfs(doc: &roxmltree::Document, ns: &str) -> Vec<CellXf> {
+    let mut xfs = Vec::new();
+    for xfs_node in doc.descendants() {
+        if xfs_node.tag_name().name() == "cellXfs" && xfs_node.tag_name().namespace() == Some(ns) {
+            for xf_node in xfs_node.children() {
+                if xf_node.tag_name().name() != "xf" { continue; }
+                let font_id = xf_node.attribute("fontId").and_then(|v| v.parse().ok()).unwrap_or(0);
+                let fill_id = xf_node.attribute("fillId").and_then(|v| v.parse().ok()).unwrap_or(0);
+                let border_id = xf_node.attribute("borderId").and_then(|v| v.parse().ok()).unwrap_or(0);
+                let num_fmt_id = xf_node.attribute("numFmtId").and_then(|v| v.parse().ok()).unwrap_or(0);
+                let mut align_h = None;
+                let mut align_v = None;
+                let mut wrap_text = false;
+                let mut indent = None;
+                let mut text_rotation = None;
+                let mut shrink_to_fit = false;
+                let mut reading_order: Option<u32> = None;
+                for child in xf_node.children() {
+                    if child.tag_name().name() == "alignment" {
+                        align_h = child.attribute("horizontal").map(|s| s.to_string());
+                        align_v = child.attribute("vertical").map(|s| s.to_string());
+                        wrap_text = child.attribute("wrapText").map(|v| v == "1" || v == "true").unwrap_or(false);
+                        indent = child.attribute("indent").and_then(|s| s.parse::<u32>().ok()).filter(|&v| v > 0);
+                        text_rotation = child.attribute("textRotation").and_then(|s| s.parse::<u32>().ok()).filter(|&v| v > 0);
+                        shrink_to_fit = child.attribute("shrinkToFit").map(|v| v == "1" || v == "true").unwrap_or(false);
+                        reading_order = child.attribute("readingOrder").and_then(|s| s.parse::<u32>().ok()).filter(|&v| v > 0);
+                    }
+                }
+                xfs.push(CellXf { font_id, fill_id, border_id, num_fmt_id, align_h, align_v, wrap_text, indent, text_rotation, shrink_to_fit, reading_order });
+            }
+            break;
+        }
+    }
+    xfs
+}
+
